返回自定义DTO 类,DTO嵌套

+++ src/main/java/com/example/px_service/domain/User.java
@@ -65,7 +65,7 @@ public class User {
     private Integer deleteAt;
     //=====set and get
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
